Added new components/pages/hooks structure
