+++ src/components/Dashboard.tsx
@@ -5,7 +5,7 @@ import { Navigate } from 'react-router-dom';
 
 const auth = getAuth();
 
-const DashboardPage = () => {
+const Dashboard = () => {
   const [user, userLoading] = useAuthState(auth);
   const [signOut, isSigningOut] = useSignOut(auth);
 
@@ -29,4 +29,4 @@ const DashboardPage = () => {
   );
 };
 
-export default DashboardPage;
+export default Dashboard;
