+++ src/hooks/useFetchTaskList.ts
@@ -1,0 +1,4 @@
+export function useFetchTaskList() {
+  // Here goes the functionality of fetching tasks from Firestore.
+  return [];
+}
