+++ src/pages/AuthPage.tsx
@@ -1,0 +1,5 @@
+import AuthForm from '../components/AuthForm';
+
+export function AuthPage() {
+  return <AuthForm />;
+}
