+++ src/pages/DashboardPage.tsx
@@ -1,0 +1,9 @@
+import Dashboard from '../components/Dashboard';
+
+/**
+ *
+ * @deprecated an should be removed later
+ */
+export function DashboardPage() {
+  return <Dashboard />;
+}
