+++ src/pages/FirebaseDemoPage.tsx
@@ -1,0 +1,9 @@
+import FirebaseDemo from '../components/FirebaseDemo';
+
+/**
+ *
+ * @deprecated an should be removed later
+ */
+export function FirebaseDemoPage() {
+  return <FirebaseDemo />;
+}
