+++ src/pages/TasksPage.tsx
@@ -1,0 +1,10 @@
+import MainLayout from '../components/layout/MainLayout';
+import { TaskList } from '../components/TaskList';
+
+export function TasksPage() {
+  return (
+    <MainLayout>
+      <TaskList />
+    </MainLayout>
+  );
+}
